pyproject.toml
[project]
name = "trading212-mcp-server"
version = "0.1.0"
description = "MCP server for Trading 212 API integration"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "Rishabh Dogra", email = "contact@kyuish.com" }]
keywords = ["mcp", "trading212", "trading", "portfolio", "stocks"]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.8.0",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
]

[project.urls]
Homepage = "https://github.com/KyuRish/trading212-mcp-server"
Repository = "https://github.com/KyuRish/trading212-mcp-server"
"Bug Tracker" = "https://github.com/KyuRish/trading212-mcp-server/issues"

[project.scripts]
trading212-mcp-server = "trading212_mcp_server.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/trading212_mcp_server"]
